+++ Code/Launcher/Launcher.h
@@ -13,6 +13,7 @@ class Launcher : public ISystemUserCallback
 	DLL m_CryAction;
 	DLL m_CryNetwork;
 	DLL m_CrySystem;
+	DLL m_CryRenderD3D10;
 	SSystemInitParams m_params;
 	GameWindow m_gameWindow;
 	CGame *m_pGame = nullptr;